+++ NEWS.md
@@ -1,5 +1,11 @@
 # dwtools 1.0.0 (in dev)
 
 * db
+  * tested on csv (default if db conns not provided)
+  * tested on SQLite, postgres (should works on any DBI driver)
+  * TODO: test ODBC
+  * TODO: couchdb
 * joinbyv
-* CJI
+* populate star schema data
+* CJI - soon will be tested and confirmed
+* timing - in dev